requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiomysql==0.3.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
courlan==1.3.2
dateparser==1.2.2
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.121.1
frozenlist==1.8.0
google-auth==2.43.0
google-genai==1.49.0
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
httpx-aiohttp==0.1.9
idna==3.11
jiter==0.12.0
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
multidict==6.7.0
numpy==2.3.4
openai==2.7.1
pandas==2.3.3
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
PyMuPDF==1.26.6
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
